[emacs] Configure `cperl-mode`.
This was copied from my previous emacs configuration and adapted to
`req-package`.  There are a few customizations to make the indentation
more like the standard Perl conventions.  We also set
`cperl-invalid-face` to `nil` to prevent annoying spaces at the ends of
lines from being displayed.  When I want those, I use `whitespace-mode`.

In the future I'd like to integrate `perlcritic` and `perltidy` into
this as well.