[emacs] Add window number utilities
We add a function `do-in-window-number`, which executes a function in a
given number window, and a function `find-file-in-window`, which is the
same as `find-file`, but within a given number window.  This last
function is bound to `C-c f` for ease of use.

Finally, we add an eshell alias so running `emacs FILENAME N` runs
`(find-file-in-window FILENAME N)`.