[emacs] Set up clang-format integration
Using the `clang-format` package on MELPA, we can integrate Emacs with
clang-format.  We can now use the `C-TAB` binding to properly format the
selected region.

For this to work, we have to make sure `cc-mode` has already been
loaded when `clang-format` is, but `cc-mode` isn't in ELPA/MELPA.
Instead, we use the normal Emacs `require` construct at init time.

Also, the `clang-format` package on MELPA wants an executable named
`clang-format`, but Debian doesn't provide one (it suffixes the version
number onto the executable).  As such, we provide one ourselves which
just dispatches to the correct version.