[ssh] Import SSH configuration
I vetted this to make sure there was no personal information.  I'm not
sure how well this configuration can move between different systems.  If
I need to port it to a different system, or if I start to add more
personal information to this, I may come back to this in the future and
refactor it into multiple files.  There doesn't seem to be good way to
`include` another file from the `ssh_config` file.

License: CC0
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>