[gpg] Update GnuPG configuration
We move this back to `.gnupg` instead of trying to mess around with
having it elsewhere, because it's hard to make sure the proper
environment variables are always set when we want to use GPG (for
example, in a cronjob).

Debian has updated its `gpg2` package to GnuPG 2.1, which requires a
separate configuration file for verifying the certificate of the
keyserver.  I'm not 100% sure why this is the case, but it looks like a
tool named `dirmngr` was introduced.  I should research this; I may want
to add more settings to this configuration file.

Instead of making `~/.gnupg` itself a symlink to the repository's
directory, we only symlink the configurations inside of it into the
repository.  This is safer than relying on a `.gitignore` in the
repository to make sure nothing private is committed and distributed.

License: CC0
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>