[mail] Import mail settings
I keep a local copy of my mail messages which I sync over IMAP with
`mbsync`.  This allows me to read my mail without an internet
connection, which is incredibly useful to me.  All the passwords are
stored in a `pass` database, and none of the servers are open relays, so
I don't have to worry about committing this configuration.

My mail server has a self-signed certificate, so I am keeping a copy in
this repo.  My hope is to switch this out with a cert signed by Let's
Encrypt when they go fully public.

When I'm ready to send a message, I use `msmtp`, which works just like
sendmail.  In the Debian package, there are a set of scripts for
maintaining a queue so I can send email offline.  This would be very,
very useful, and I intend to look into this in the future.

License: CC0
Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>